Remove Elasticsearch6QueryCompiler and make Elasticsearch7QueryCompiler the base implementation
