+++ wagtail/search/backends/elasticsearch6.py
@@ -1,13 +1,8 @@
-import json
-
-from django.db import DEFAULT_DB_ALIAS, models
-from django.db.models.sql import Query
-from django.db.models.sql.constants import MULTI
+from django.db import models
 from elasticsearch import NotFoundError
 from elasticsearch.helpers import bulk
 
 from wagtail.search.backends.base import (
-    BaseSearchQueryCompiler,
     get_model_root,
 )
 from wagtail.search.index import (
@@ -17,9 +12,7 @@ from wagtail.search.index import (
     RelatedFields,
     SearchField,
     class_is_indexed,
-    get_indexed_models,
 )
-from wagtail.search.query import And, Boost, Fuzzy, MatchAll, Not, Or, Phrase, PlainText
 
 
 class Field:
@@ -415,353 +408,3 @@ class Elasticsearch6Index:
 
         # Create new index
         self.put()
-
-
-class Elasticsearch6SearchQueryCompiler(BaseSearchQueryCompiler):
-    mapping_class = Elasticsearch6Mapping
-    DEFAULT_OPERATOR = "or"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.mapping = self.mapping_class(self.queryset.model)
-
-        # Convert field names into index column names
-        if self.fields:
-            self.remapped_fields = []
-            searchable_fields = {
-                f.field_name: f
-                for f in self.queryset.model.get_searchable_search_fields()
-            }
-            for field_name in self.fields:
-                if field_name in searchable_fields:
-                    field_name = self.mapping.get_field_column_name(
-                        searchable_fields[field_name]
-                    )
-
-                self.remapped_fields.append(Field(field_name))
-        else:
-            self.remapped_fields = [Field(self.mapping.all_field_name)]
-
-        models = get_indexed_models()
-        unique_boosts = set()
-        for model in models:
-            for field in model.get_searchable_search_fields():
-                if field.boost:
-                    unique_boosts.add(float(field.boost))
-
-        self.remapped_fields.extend(
-            [
-                Field(self.mapping.get_boost_field_name(boost), boost)
-                for boost in unique_boosts
-            ]
-        )
-
-    def get_boosted_fields(self, fields):
-        boosted_fields = []
-        if not isinstance(fields, list):
-            fields = [fields]
-        for field in fields:
-            if field.boost != 1:
-                boosted_fields.append(f"{field.field_name}^{field.boost}")
-            else:
-                boosted_fields.append(field.field_name)
-        return boosted_fields
-
-    def _process_lookup(self, field, lookup, value):
-        column_name = self.mapping.get_field_column_name(field)
-
-        if lookup == "exact":
-            if value is None:
-                return {
-                    "missing": {
-                        "field": column_name,
-                    }
-                }
-            else:
-                return {
-                    "term": {
-                        column_name: value,
-                    }
-                }
-
-        if lookup == "isnull":
-            query = {
-                "exists": {
-                    "field": column_name,
-                }
-            }
-
-            if value:
-                query = {"bool": {"mustNot": query}}
-
-            return query
-
-        if lookup in ["startswith", "prefix"]:
-            return {
-                "prefix": {
-                    column_name: value,
-                }
-            }
-
-        if lookup in ["gt", "gte", "lt", "lte"]:
-            return {
-                "range": {
-                    column_name: {
-                        lookup: value,
-                    }
-                }
-            }
-
-        if lookup == "range":
-            lower, upper = value
-
-            return {
-                "range": {
-                    column_name: {
-                        "gte": lower,
-                        "lte": upper,
-                    }
-                }
-            }
-
-        if lookup == "in":
-            if isinstance(value, Query):
-                db_alias = self.queryset._db or DEFAULT_DB_ALIAS
-                resultset = value.get_compiler(db_alias).execute_sql(result_type=MULTI)
-                value = [row[0] for chunk in resultset for row in chunk]
-
-            elif not isinstance(value, list):
-                value = list(value)
-            return {
-                "terms": {
-                    column_name: value,
-                }
-            }
-
-    def _connect_filters(self, filters, connector, negated):
-        if filters:
-            if len(filters) == 1:
-                filter_out = filters[0]
-            elif connector == "AND":
-                filter_out = {
-                    "bool": {"must": [fil for fil in filters if fil is not None]}
-                }
-            elif connector == "OR":
-                filter_out = {
-                    "bool": {"should": [fil for fil in filters if fil is not None]}
-                }
-
-            if negated:
-                filter_out = {"bool": {"mustNot": filter_out}}
-
-            return filter_out
-
-    def _compile_plaintext_query(self, query, fields, boost=1.0):
-        fields = self.get_boosted_fields(fields)
-        match_query = {"query": query.query_string}
-
-        if query.operator != "or":
-            match_query["operator"] = query.operator
-
-        if boost != 1.0:
-            match_query["boost"] = boost
-
-        if len(fields) == 1:
-            return {"match": {fields[0]: match_query}}
-        else:
-            match_query["fields"] = fields
-
-            return {"multi_match": match_query}
-
-    def _compile_fuzzy_query(self, query, fields):
-        if len(fields) == 1:
-            return {
-                "match": {
-                    fields[0]: {
-                        "query": query.query_string,
-                        "fuzziness": "AUTO",
-                    }
-                }
-            }
-        return {
-            "multi_match": {
-                "query": query.query_string,
-                "fields": self.get_boosted_fields(fields),
-                "fuzziness": "AUTO",
-            }
-        }
-
-    def _compile_phrase_query(self, query, fields):
-        fields = self.get_boosted_fields(fields)
-        if len(fields) == 1:
-            return {"match_phrase": {fields[0]: query.query_string}}
-        else:
-            return {
-                "multi_match": {
-                    "query": query.query_string,
-                    "fields": fields,
-                    "type": "phrase",
-                }
-            }
-
-    def _compile_query(self, query, field, boost=1.0):
-        if isinstance(query, MatchAll):
-            match_all_query = {}
-
-            if boost != 1.0:
-                match_all_query["boost"] = boost
-
-            return {"match_all": match_all_query}
-
-        elif isinstance(query, And):
-            return {
-                "bool": {
-                    "must": [
-                        self._compile_query(child_query, field, boost)
-                        for child_query in query.subqueries
-                    ]
-                }
-            }
-
-        elif isinstance(query, Or):
-            return {
-                "bool": {
-                    "should": [
-                        self._compile_query(child_query, field, boost)
-                        for child_query in query.subqueries
-                    ]
-                }
-            }
-
-        elif isinstance(query, Not):
-            return {
-                "bool": {"mustNot": self._compile_query(query.subquery, field, boost)}
-            }
-
-        elif isinstance(query, PlainText):
-            return self._compile_plaintext_query(query, [field], boost)
-
-        elif isinstance(query, Fuzzy):
-            return self._compile_fuzzy_query(query, [field])
-
-        elif isinstance(query, Phrase):
-            return self._compile_phrase_query(query, [field])
-
-        elif isinstance(query, Boost):
-            return self._compile_query(query.subquery, field, boost * query.boost)
-
-        else:
-            raise NotImplementedError(
-                "`%s` is not supported by the Elasticsearch search backend."
-                % query.__class__.__name__
-            )
-
-    def get_inner_query(self):
-        if self.remapped_fields:
-            fields = self.remapped_fields
-        else:
-            fields = [self.mapping.all_field_name]
-
-        if len(fields) == 0:
-            # No fields. Return a query that'll match nothing
-            return {"bool": {"mustNot": {"match_all": {}}}}
-
-        # Handle MatchAll and PlainText separately as they were supported
-        # before "search query classes" was implemented and we'd like to
-        # keep the query the same as before
-        if isinstance(self.query, MatchAll):
-            return {"match_all": {}}
-
-        elif isinstance(self.query, PlainText):
-            return self._compile_plaintext_query(self.query, fields)
-
-        elif isinstance(self.query, Phrase):
-            return self._compile_phrase_query(self.query, fields)
-
-        elif isinstance(self.query, Fuzzy):
-            return self._compile_fuzzy_query(self.query, fields)
-
-        elif isinstance(self.query, Not):
-            return {
-                "bool": {
-                    "mustNot": [
-                        self._compile_query(self.query.subquery, field)
-                        for field in fields
-                    ]
-                }
-            }
-
-        else:
-            if len(fields) == 1:
-                return self._compile_query(self.query, fields[0])
-            else:
-                # Compile a query for each field then combine with disjunction
-                # max (or operator which takes the max score out of each of the
-                # field queries)
-                field_queries = []
-                for field in fields:
-                    field_queries.append(self._compile_query(self.query, field))
-
-                return {"dis_max": {"queries": field_queries}}
-
-    def get_content_type_filter(self):
-        # Query content_type using a "match" query. See comment in
-        # Elasticsearch6Mapping.get_document for more details
-        content_type = self.mapping_class(self.queryset.model).get_content_type()
-
-        return {"match": {"content_type": content_type}}
-
-    def get_filters(self):
-        # Filter by content type
-        filters = [self.get_content_type_filter()]
-
-        # Apply filters from queryset
-        queryset_filters = self._get_filters_from_queryset()
-        if queryset_filters:
-            filters.append(queryset_filters)
-
-        return filters
-
-    def get_query(self):
-        inner_query = self.get_inner_query()
-        filters = self.get_filters()
-
-        if len(filters) == 1:
-            return {
-                "bool": {
-                    "must": inner_query,
-                    "filter": filters[0],
-                }
-            }
-        elif len(filters) > 1:
-            return {
-                "bool": {
-                    "must": inner_query,
-                    "filter": filters,
-                }
-            }
-        else:
-            return inner_query
-
-    def get_sort(self):
-        # Ordering by relevance is the default in Elasticsearch
-        if self.order_by_relevance:
-            return
-
-        # Get queryset and make sure its ordered
-        if self.queryset.ordered:
-            sort = []
-
-            for reverse, field in self._get_order_by():
-                column_name = self.mapping.get_field_column_name(field)
-
-                sort.append({column_name: "desc" if reverse else "asc"})
-
-            return sort
-
-        else:
-            # Order by pk field
-            return ["pk"]
-
-    def __repr__(self):
-        return json.dumps(self.get_query())
